requirements.test.txt
-r requirements.txt
pytest-django==4.9.0
